# Optimize range iteration