--- a/lib/liquid.rb
+++ b/lib/liquid.rb
@@ -74,6 +74,7 @@ require 'liquid/condition'
 require 'liquid/utils'
 require 'liquid/tokenizer'
 require 'liquid/parse_context'
+require 'liquid/reversable_range'
 
 # Load all the tags of the standard library
 #