For loops support a range syntax for iterating a fixed number of times,
which looks like this.

```liquid
{% for i in range (1..100) %}
        ...
{% endfor %}
```

Previously, we converted these ranges to arrays using `to_a`, which
initialized an array containing each number in the range. Since all we
use these ranges for is iteration, this is far less efficient than
simply keeping the ranges as ranges.

Doing this means that iterating over ranges now takes O(1) rather than O(n)
memory. 

* Remove to_a cast on ranges
* Add helper method to check for empty ranges
* Add custom range-specific slicing logic

### Benchmarks

**TLDR**
CPU: ~10% speedup in simple cases, →∞% speedup in pathological cases.
Memory: 7% less usage in simple cases, →∞% less memory in pathological cases

This comes from a change of  `O(n)` -> `O(1)` in terms of memory usage.

**Simple iteration, CPU**
<details>
<summary>Code</summary>

```ruby
require 'benchmark/ips'

require 'liquid.rb'

Benchmark.ips do |x|
  x.time = 10
  x.warmup = 5

  puts
  puts "Running benchmark for #{x.time} seconds (with #{x.warmup} seconds warmup)."
  puts

  small_range_template = <<~LIQUID
  {% for _each in (1..100) %}
    Hello world
  {% endfor %}
  LIQUID
  small_range = Liquid::Template.parse(small_range_template)
  x.report("small range") do
    small_range.render!({})
  end

  large_range_template = <<~LIQUID
  {% for _each in (1..1000000) %}
    Hello world
  {% endfor %}
  LIQUID
  large_range = Liquid::Template.parse(large_range_template)
  x.report("large range") do
    large_range.render!({})
  end

  early_exit_template = <<~LIQUID
  {% for _each in (1..1000000) %}
    {% break %}
  {% endfor %}
  LIQUID
  early_exit = Liquid::Template.parse(early_exit_template)
  x.report("early exit") do
    early_exit.render!({})
  end
end
```

</details>

<details>
<summary>Before</summary>

```
Running benchmark for 10 seconds (with 5 seconds warmup).

Warming up --------------------------------------
         small range   796.000  i/100ms
         large range     1.000  i/100ms
          early exit     1.000  i/100ms
Calculating -------------------------------------
         small range      8.662k (± 2.7%) i/s -     86.764k in  10.023757s
         large range      0.930  (± 0.0%) i/s -     10.000  in  10.749791s
          early exit     10.655  (± 0.0%) i/s -    107.000  in  10.048596s
```
</details>
<details>
<summary>After</summary>

```
Running benchmark for 10 seconds (with 5 seconds warmup).

Warming up --------------------------------------
         small range   944.000  i/100ms
         large range     1.000  i/100ms
          early exit     9.465k i/100ms
Calculating -------------------------------------
         small range      9.589k (± 1.9%) i/s -     96.288k in  10.045111s
         large range      1.051  (± 0.0%) i/s -     11.000  in  10.466125s
          early exit     99.545k (± 4.1%) i/s -      1.003M in  10.097012s
```
</details>

**Simple iteration, Memory**

<details>
<summary>Code</summary>

```ruby
require 'liquid.rb'
require 'memory_profiler'

def profile(phase, &block)
  puts
  puts "#{phase}:"
  puts

  report = MemoryProfiler.report(&block)

  report.pretty_print(
    color_output: true,
    scale_bytes: true,
    detailed_report: true
  )
end

small_range_template = <<~LIQUID
{% for _each in (1..100) %}
  Hello world
{% endfor %}
LIQUID
small_range = Liquid::Template.parse(small_range_template)
profile("small range") do
  small_range.render!({})
end

large_range_template = <<~LIQUID
{% for _each in (1..1000000) %}
  Hello world
{% endfor %}
LIQUID
large_range = Liquid::Template.parse(large_range_template)
profile("large range") do
  large_range.render!({})
end

early_exit_template = <<~LIQUID
{% for _each in (1..1000000) %}
  {% break %}
{% endfor %}
LIQUID
early_exit = Liquid::Template.parse(early_exit_template)
profile("early exit") do
  early_exit.render!({})
end
```
</details>

<details>
<summary>Before</summary>

```

small range:

Total allocated: 28.26 kB (226 objects)
Total retained:  1.05 kB (7 objects)

allocated memory by gem
-----------------------------------
  27.99 kB  liquid/lib
   272.0 B  other

allocated memory by file
-----------------------------------
  22.38 kB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/block_body.rb
   3.03 kB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb
   1.06 kB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/utils.rb
   920.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb
   592.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb
   272.0 B  for_range_memory_profile.rb

allocated memory by location
-----------------------------------
  18.34 kB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/block_body.rb:97
   4.04 kB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/block_body.rb:70
   1.58 kB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:157
   1.06 kB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:136
   1.06 kB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/utils.rb:27
   232.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:137
   232.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:122
   232.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:138
   232.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:142
   232.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:146
   232.0 B  for_range_memory_profile.rb:25
   184.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:183
    80.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:19
    80.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:32
    80.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:160
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:139
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:20
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:22
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:35
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:36
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:148
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:154
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:221
    40.0 B  for_range_memory_profile.rb:24

allocated memory by class
-----------------------------------
   20.0 kB  String
   6.57 kB  Array
   1.39 kB  Hash
   144.0 B  Liquid::Context
    80.0 B  Liquid::ForloopDrop
    80.0 B  Proc

allocated objects by gem
-----------------------------------
       224  liquid/lib
         2  other

allocated objects by file
-----------------------------------
       202  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/block_body.rb
         9  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb
         6  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb
         6  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb
         2  for_range_memory_profile.rb
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/utils.rb

allocated objects by location
-----------------------------------
       101  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/block_body.rb:70
       101  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/block_body.rb:97
         2  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:19
         2  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:183
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:137
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:139
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:20
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:22
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:32
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:35
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:36
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:122
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:136
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:148
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:154
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:157
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:160
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:138
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:142
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:146
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:221
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/utils.rb:27
         1  for_range_memory_profile.rb:24
         1  for_range_memory_profile.rb:25

allocated objects by class
-----------------------------------
       113  Array
       104  String
         6  Hash
         1  Liquid::Context
         1  Liquid::ForloopDrop
         1  Proc

retained memory by gem
-----------------------------------
   1.05 kB  liquid/lib

retained memory by file
-----------------------------------
   696.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb
   312.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb

retained memory by location
-----------------------------------
   232.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:122
   232.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:138
   232.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:142
   232.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:146
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:22
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:148
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:154

retained memory by class
-----------------------------------
   928.0 B  Hash
    80.0 B  Array
    40.0 B  String

retained objects by gem
-----------------------------------
         7  liquid/lib

retained objects by file
-----------------------------------
         3  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb
         3  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb

retained objects by location
-----------------------------------
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:22
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:122
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:148
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:154
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:138
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:142
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:146

retained objects by class
-----------------------------------
         4  Hash
         2  Array
         1  String


Allocated String Report
-----------------------------------
       100  "\n  Hello world\n"
       100  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/block_body.rb:97

         1  "\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  He"
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/block_body.rb:97

         1  "\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  Hello world\n\n  He"
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:157

         1  "_each"
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:139

         1  "_each-(1..100)"
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:148


Retained String Report
-----------------------------------
         1  "_each-(1..100)"
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:148


large range:

Total allocated: 271.44 MB (2000026 objects)
Total retained:  1.05 kB (7 objects)

allocated memory by gem
-----------------------------------
 271.44 MB  liquid/lib
   272.0 B  other

allocated memory by file
-----------------------------------
  223.0 MB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/block_body.rb
   36.8 MB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb
  11.64 MB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/utils.rb
   912.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb
   592.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb
   272.0 B  for_range_memory_profile.rb

allocated memory by location
-----------------------------------
  183.0 MB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/block_body.rb:97
   40.0 MB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/block_body.rb:70
  25.17 MB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:157
  11.64 MB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:136
  11.64 MB  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/utils.rb:27
   232.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:137
   232.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:122
   232.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:138
   232.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:142
   232.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:146
   232.0 B  for_range_memory_profile.rb:35
   176.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:183
    80.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:19
    80.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:32
    72.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:160
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:139
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:20
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:22
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:35
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:36
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:148
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:154
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:221
    40.0 B  for_range_memory_profile.rb:34

allocated memory by class
-----------------------------------
 208.17 MB  String
  63.27 MB  Array
   1.39 kB  Hash
   136.0 B  Liquid::Context
    80.0 B  Proc
    72.0 B  Liquid::ForloopDrop

allocated objects by gem
-----------------------------------
   2000024  liquid/lib
         2  other

allocated objects by file
-----------------------------------
   2000002  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/block_body.rb
         9  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb
         6  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb
         6  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb
         2  for_range_memory_profile.rb
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/utils.rb

allocated objects by location
-----------------------------------
   1000001  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/block_body.rb:70
   1000001  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/block_body.rb:97
         2  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:19
         2  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:183
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:137
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:139
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:20
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:22
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:32
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:35
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb:36
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:122
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:136
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:148
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:154
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:157
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb:160
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:138
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:142
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:146
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb:221
         1  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/utils.rb:27
         1  for_range_memory_profile.rb:34
         1  for_range_memory_profile.rb:35

allocated objects by class
-----------------------------------
   1000013  Array
   1000004  String
         6  Hash
         1  Liquid::Context
         1  Liquid::ForloopDrop
         1  Proc

retained memory by gem
-----------------------------------
   1.05 kB  liquid/lib

retained memory by file
-----------------------------------
   696.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/template.rb
   312.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/tags/for.rb
    40.0 B  /Users/samdoiron/src/github.com/Shopify/liquid/lib/liquid/context.rb

retained memory by location
…