@samdoiron 2019-06-25
Unfortunately our integration benchmarks to not use ranges, so there was no change there. However, this should see a reasonable speedup / memory savings in practical usage.