@ashmaroli 2019-06-25
Why the change from `#length` to `#size`? The two methods are aliases of each other..