@samdoiron 2019-06-25
@ashmaroli They are aliases in many cases, but `length` is not defined on ranges, whereas `size` is.