@ashmaroli 2019-06-25
> but `length` is not defined on ranges

Ha! T.I.L.
Thanks :smiley: