@ashmaroli 2019-06-25
@samdoiron You can improve the readability of the reports from MemoryProfiler by passing `color_output: false` to `Report#pretty_print`.. The escape sequences are too noisy..