@samdoiron 2019-07-05
A quick note, since I haven't been able to focus on this the last couple of days: this will require a concurrent change in Liquid C, which directly generates Range objects. I hope to have a separate PR there to update this behaviour soon.