@shopmike 2019-08-29
@samdoiron I just opened up a PR which is a possible match to this for the liquid-c side. https://github.com/Shopify/liquid-c/pull/48