@shopmike 2019-08-29
I've just realised you already did one https://github.com/Shopify/liquid-c/pull/46/files. Oh well I guess either implementation is an option