@shopmike 2019-09-02
@pushrax can we get your feedback on this and the approach for liquid-c.