@shopmike 2019-09-16
I'm assuming this would also fix https://github.com/Shopify/liquid/pull/1071