@samdoiron 2019-09-17
Yes, it would.